+++ src/splat/segtypes/segment.py
@@ -187,11 +187,10 @@ class Segment:
         return None
 
     @staticmethod
-    def parse_segment_bss_contains_common(segment: Union[dict, list]) -> bool:
+    def parse_segment_bss_contains_common(segment: Union[dict, list], default: bool) -> bool:
         if isinstance(segment, dict) and "bss_contains_common" in segment:
             return bool(segment["bss_contains_common"])
-        else:
-            return False
+        return default
 
     @staticmethod
     def parse_linker_section_order(yaml: Union[dict, list]) -> Optional[str]:
@@ -272,7 +271,7 @@ class Segment:
 
         self.warnings: List[str] = []
         self.did_run = False
-        self.bss_contains_common = Segment.parse_segment_bss_contains_common(yaml)
+        self.bss_contains_common = Segment.parse_segment_bss_contains_common(yaml, options.opts.ld_bss_contains_common)
 
         # For segments which are not in the usual VRAM segment space, like N64's IPL3 which lives in 0xA4...
         self.special_vram_segment: bool = False
@@ -341,7 +340,7 @@ class Segment:
         )
         ret.file_path = Segment.parse_segment_file_path(yaml)
 
-        ret.bss_contains_common = Segment.parse_segment_bss_contains_common(yaml)
+        ret.bss_contains_common = Segment.parse_segment_bss_contains_common(yaml, options.opts.ld_bss_contains_common)
 
         ret.given_follows_vram = parse_segment_follows_vram(yaml)
         ret.given_vram_symbol = parse_segment_vram_symbol(yaml)