+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ### 0.21.6
 
-* Fix `bss_contains_common` option not being passed to "auto all" inserted sections
+* Fix `bss_contains_common` option not being passed to "auto all" inserted sections.
+* New yaml option: `ld_bss_contains_common`
+  * Sets the default option for the `bss_contains_common` attribute of all segments.
 
 ### 0.21.5